requirements.txt
absl-py==2.1.0
astunparse==1.6.3
audioread==3.0.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
mtcnn==0.1.1
mysql-connector-python==9.0.0
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==4.25.4
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
requests==2.32.3
Resemblyzer==0.1.4
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
shiboken6==6.7.2
six==1.16.0
sounddevice==0.4.7
soundfile==0.12.1
soxr==0.4.0
sympy==1.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.4.0
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.2
webrtcvad==2.0.10
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
